Build for 64bit only and set default --to folder. - no min version on macosx. - Add 'x64' compiler detection in bootstrap.mak   fixes https://ghosthub.corp.blizzard.net/blizzard/premake/issues/102

+++ src/host/criteria_matches.c
@@ -52,7 +52,7 @@ int criteria_compile(lua_State* L)
 	lua_setmetatable(L, -2);
 
 	/* create array to hold the incoming list of patterns */
-	n = lua_rawlen(L, 1);
+	n = (int)lua_rawlen(L, 1);
 	patterns->n = n;
 	patterns->pattern = (struct Pattern*)malloc(sizeof(struct Pattern) * n);
 
@@ -82,7 +82,7 @@ int criteria_compilePattern(lua_State* L, struct Pattern* pattern)
 	int i, n;
 
 	/* create array to hold the incoming list of words */
-	n = lua_rawlen(L, -1);
+	n = (int)lua_rawlen(L, -1);
 	pattern->n = n;
 	pattern->word = (struct Word*)malloc(sizeof(struct Word) * n);
 	pattern->matchesFiles = 0;
@@ -174,11 +174,11 @@ static int match(lua_State* L, const char* value, struct Word* word)
 static int testValue(lua_State* L, struct Word* word)
 {
 	const char* value;
-	size_t i, n;
+	int i, n;
 	int result;
 
 	if (lua_istable(L, -1)) {
-		n = lua_rawlen(L, -1);
+		n = (int)lua_rawlen(L, -1);
 		for (i = 1; i <= n; ++i) {
 			lua_rawgeti(L, -1, i);
 			result = testValue(L, word);
